requirements.txt
streamlit
langchain
langchain_community
ollama
plotly
networkx
streamlit_authenticator
pyyaml
neo4j
regex
openpyxl
streamlit_option_menu
streamlit-table-select-cell
pyvis
psycopg2
openpyxl
xlrd
scipy